move dashboard layout files to user folder
