+++ frontend/src/user/layout/UserLayout.tsx
@@ -1,6 +1,6 @@
 import {Outlet, useNavigate} from "react-router";
 import {useEffect} from "react";
-import {getSession} from "../utils";
+import {getSession} from "../../utils";
 import TopBar from "./TopBar";
 
 export default function UserLayout() {
